refactor-identifier: move BadNameError to error.rs
