+++ src/lib/identifiers.rs
@@ -5,6 +5,7 @@ use std::{fmt, str::FromStr, ops};
 use serde::{Serialize, Deserialize};
 use uuid::Uuid;
 
+use crate::error::BadNameError;
 #[allow(unused)]
 use crate::models::stopwatch::Stopwatch; // for see also documentation
 
@@ -52,18 +53,6 @@ impl From<Uuid> for UniqueId {
         Self::new(bytes)
     }
 }
-
-/// If a name starts with '@'.
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct BadNameError;
-
-impl fmt::Display for BadNameError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "BadNameError")
-    }
-}
-
-impl std::error::Error for BadNameError { }
 
 #[derive(Clone, Debug, Default, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub struct Name(String);
